# api/openapi.yaml
openapi: 3.1.0
info:
  title: sshc Local API
  version: 0.1.0
paths:
...
  /api/v1/known-hosts:
    get:
      operationId: listKnownHosts
      parameters:
        - name: query
          in: query
          required: false
          schema: { type: string, maxLength: 255 }
      responses:
        "200":
          description: Matching known_hosts entries
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KnownHostsResponse" }
        "401": { $ref: "#/components/responses/Problem" }
  /api/v1/known-hosts/delete:
    post:
      operationId: deleteKnownHosts
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KnownHostsDeleteRequest" }
      responses:
        "200":
          description: Entries removed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KnownHostsChangeResponse" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/known-hosts/scan:
    post:
      operationId: scanKnownHosts
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KnownHostsScanRequest" }
      responses:
        "200":
          description: Unverified candidates
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KnownHostsScanResponse" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
  /api/v1/known-hosts/add:
    post:
      operationId: addKnownHost
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KnownHostsAddRequest" }
      responses:
        "200":
          description: Entry added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KnownHostsChangeResponse" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
...
  /api/v1/login-item:
    get:
      operationId: getLoginItem
      responses:
        "200":
          description: Whether this application starts at login
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginItem" }
        "401": { $ref: "#/components/responses/Problem" }
    put:
      operationId: setLoginItem
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginItem" }
      responses:
        "200":
          description: The setting after the change
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginItem" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
...
components:
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        detail: { type: string }
        path: { type: string }
        line: { type: integer }
        column: { type: integer }
        diagnostics: { type: array, items: { $ref: "#/components/schemas/Diagnostic" } }
        conflict: { $ref: "#/components/schemas/ConflictReport" }
...
    LoginItem:
      type: object
      additionalProperties: false
      required: [enabled, supported]
      properties:
        # ユーザーが求めない限りオフ。保存済みのあらゆる秘密の鍵を握る
        # バックグラウンドプロセスは、他人に代わって勝手に用意してよいものではない。
        enabled: { type: boolean }
        # このビルドが自分自身のパスを解決できないとき false になる。launchd の
        # エージェントが名指ししなければならないのが、そのパスである。
        supported: { type: boolean }
...
    KnownHostsResponse:
      type: object
      additionalProperties: false
      required: [path, entries]
      properties:
        path: { type: string }
        entries:
          type: array
          items: { $ref: "#/components/schemas/KnownHostEntry" }
    KnownHostEntry:
      type: object
      additionalProperties: false
      required: [line, digest, marker, hosts, hashed, keyType, fingerprint, comment]
      properties:
        line: { type: integer }
        digest: { type: string }
        marker: { type: string }
        hosts:
          type: array
          items: { type: string }
        hashed: { type: boolean }
        keyType: { type: string }
        fingerprint: { type: string }
        comment: { type: string }
    KnownHostTarget:
      type: object
      additionalProperties: false
      required: [line, digest]
      properties:
        line: { type: integer }
        digest: { type: string, minLength: 64, maxLength: 64 }
    KnownHostsDeleteRequest:
      type: object
      additionalProperties: false
      required: [entries]
      properties:
        entries:
          type: array
          items: { $ref: "#/components/schemas/KnownHostTarget" }
    KnownHostsChangeResponse:
      type: object
      additionalProperties: false
      required: [changed, transactionId]
      properties:
        changed: { type: boolean }
        transactionId: { type: string }
    KnownHostsScanRequest:
      type: object
      additionalProperties: false
      required: [host, port]
      properties:
        host: { type: string, minLength: 1, maxLength: 255 }
        port: { type: integer }
    KnownHostsScanResponse:
      type: object
      additionalProperties: false
      required: [notice, candidates]
      properties:
        notice: { type: string }
        candidates:
          type: array
          items: { $ref: "#/components/schemas/KnownHostCandidate" }
    KnownHostCandidate:
      type: object
      additionalProperties: false
      required: [host, port, keyType, key, fingerprint, verified]
      properties:
        host: { type: string }
        port: { type: integer }
        keyType: { type: string }
        key: { type: string }
        fingerprint: { type: string }
        verified: { type: boolean }
    KnownHostsAddRequest:
      type: object
      additionalProperties: false
      required: [host, port, keyType, key, expectedFingerprint, acknowledged]
      properties:
        host: { type: string, minLength: 1, maxLength: 255 }
        port: { type: integer }
        keyType: { type: string, minLength: 1, maxLength: 64 }
        key: { type: string, minLength: 1, maxLength: 4096 }
        expectedFingerprint: { type: string, maxLength: 128 }
        acknowledged: { type: boolean }
...
    Diagnostic:
      type: object
      additionalProperties: false
      required: [severity, code]
      properties:
        severity: { type: string }
        code: { type: string }
        path: { type: string }
        absolute: { type: string }
        external: { type: boolean }
        line: { type: integer }
        detail: { type: string }
...
    DiffLine:
      type: object
      additionalProperties: false
      required: [op, text]
      properties:
        op: { type: string }
        text: { type: string }
        oldLine: { type: integer }
        newLine: { type: integer }
...
    ConflictReport:
      type: object
      additionalProperties: false
      required: [path, externalChange, localChange]
      properties:
        path: { type: string }
        baseDigest: { type: string }
        diskDigest: { type: string }
        externalChange: { type: array, items: { $ref: "#/components/schemas/DiffLine" } }
        localChange: { type: array, items: { $ref: "#/components/schemas/DiffLine" } }
...
  responses:
    Problem:
      description: Request rejected
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }